refactor: extract useSdkLifecycle and simplify wallet provider
